Add osprofiler config options to generated reference
osprofiler is optional, but if it's installed we'll
load up the configuration options from the library,
but they weren't in the generated config sample so
people would have to find the osprofiler docs, or
worse the code, to figure out how to configure it.

This simply adds the osprofiler config options to the
nova config sample, which will also show up in the
config reference docs.

Change-Id: I28d35165ed77487cd49d560fb1eda4f1d640734e
Closes-Bug: #1774208