Remove nova dependencies from test_resource_provider
Now, none of the tests in test_resource_provider.py require
anything from the nova base test class, so testtools.TestCase is
used instead.

Also, the request context provided by placement (intead of nova)
is more appropriate for these tests (and doesn't require a CONF)

Because the DatabasePoisonFixture is no longer present we no longer
need the mock in AllocationTest that guards against it.

Change-Id: I62d348e4551cce7a93337e32eb540574d8291bd2