Do not use nova.test in placement.test_util
These tests are very straightforward and do not require any of the
setup and fixtures that nova.test.NoDBTestCase supplies. Also, in the
future of placement extraction we don't want the dependecy.

Note that when placement is extracted, we'll need to create a local
uuidsentinel. That's easiest to do at the time of extraction.

Partially implements blueprint placement-extract

Change-Id: Ia18442d2afcaca51203104f790072eaa77b3bbac