Clarify placement DB schema migration
This just clarifies in the release note for the optional
placement database that the database itself is not created
when running "nova-manage api_db sync", but rather the
database schema is created. This is important since a
non-trivial number of people over the years have thought
that the db sync commands actually create a database, which
they do not.

Change-Id: Ie6c3a5dc61a288935829276cc72f7f7563e20420