[placement] Add gabbi coverage for an inventory change
There's a special case of changing inventories that we didn't have
gabbit coverage for: For an existing resource provider, we allow its
inventory to be changed to have a lower total than existing allocations
(See commit 09627f2a0bd). There was no gabbi coverage for that case.
This adds it.

Change-Id: I34e104a766d4578f83a455b4b900ea50c9a6ec0f