Initial commit of the extracted Placement repo
Ran the filter_git_history.sh script from oslo.tools in order to
extract only the files (and their git history) that are part of the
placement project. The script is a little loose in what it removes, so
that if there is a commit that contains both placement and non-placement
files, all are retained. Those extra files have also been removed in
this commit.