DNM: Make tox -efunctional work in a hacked up collection
This required a fair bit of work, quite a lot of files needed to be
added and other tweaks. I present it here not so we can merge these
changes but so we can use them to come up with a more ordered approach
to getting some clean commits, starting from the best base.

The openstack in the test hierarchy was removed

To make pbr and setup.py work properly

* setup.cfg
* README.rst

were added. After that we needed

* placement/conf/* copied from nova with some adjustments, but more more
  can be done
* some __init__.py files here and there to make things into packages
* a stubbed out Database fixture in placement/tests/fixtures.py from nova
* a stubbed out policy fixture in placement/tests/policy_fixture.py

Then many from nova.* imports where change to placement.

from __future__ import absolute_import was added in several places to
deal with module name collisions upsetting python 2.7 (tests started
passing in 3.6 much sooner than 2.7)

To see this work pull the branch and 'tox -efunctional', but more
relevant: look at the diff and consider our options. There are also
notes on https://etherpad.openstack.org/p/placement-extraction-file-notes

Note that nothing (as yet) has been done with the unit tests.