Demonstrate working placement unit tests
The changes here are mostly to do with adjusting imports and modifying
mock paths, but there are a few less mechanical changes:

* .stestr.conf was needed for the unit test tox rules to be okay
* the tempdir() context manager from nova.utils was copied into
  test_policy and slightly adjusted to use non-global conf
* Some nova references in some functional tests were found during a grep
  and cleaned

Again, the intention here is not for this to merge, but to provide us
with clues for how to do a clean and incremental set of commits when
doing it for real.

Most of the real work that got this stuff to work was done in the
previous commit.